requirements.txt
PySide6
QtAwesome
rembg
onnxruntime
onnxruntime-gpu
Pillow
requests
numpy